moved components to folders
